+++ src/main/java/blusunrize/immersiveengineering/common/crafting/PotionRecipeGenerators.java
@@ -9,7 +9,13 @@
 package blusunrize.immersiveengineering.common.crafting;
 
 import blusunrize.immersiveengineering.api.crafting.*;
+import blusunrize.immersiveengineering.api.tool.BulletHandler;
+import blusunrize.immersiveengineering.common.gui.IESlot.Bullet;
+import blusunrize.immersiveengineering.common.items.BulletItem;
+import blusunrize.immersiveengineering.common.register.IEItems;
+import blusunrize.immersiveengineering.common.register.IEItems.Weapons;
 import blusunrize.immersiveengineering.common.util.IELogger;
+import blusunrize.immersiveengineering.common.util.ItemNBTHelper;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -17,6 +23,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.alchemy.Potion;
 import net.minecraft.world.item.alchemy.PotionUtils;
 import net.minecraft.world.item.alchemy.Potions;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.neoforged.neoforge.fluids.FluidType;
 
 import java.util.*;
@@ -44,25 +51,41 @@ public class PotionRecipeGenerators
 
 	public static List<BottlingMachineRecipe> getPotionBottlingRecipes()
 	{
-		Map<Potion, BottlingMachineRecipe> recipes = new HashMap<>();
-		Function<Potion, BottlingMachineRecipe> toRecipe = potion -> new BottlingMachineRecipe(
+		Map<Potion, BottlingMachineRecipe> bottleRecipes = new HashMap<>();
+		Function<Potion, BottlingMachineRecipe> toBottleRecipe = potion -> new BottlingMachineRecipe(
 				new TagOutputList(new TagOutput(PotionUtils.setPotion(new ItemStack(Items.POTION), potion))),
 				IngredientWithSize.of(new ItemStack(Items.GLASS_BOTTLE)),
 				getFluidTagForType(potion, 250)
 		);
+
+		Map<Potion, BottlingMachineRecipe> bulletRecipes = new HashMap<>();
+		Function<Potion, BottlingMachineRecipe> toBulletRecipe = potion -> {
+			ItemStack bulletStack = BulletHandler.getBulletStack(BulletItem.POTION);
+			ItemNBTHelper.setItemStack(bulletStack, "potion", PotionUtils.setPotion(new ItemStack(Items.POTION), potion));
+			return new BottlingMachineRecipe(
+					new TagOutputList(new TagOutput(bulletStack)),
+					new IngredientWithSize(Ingredient.of(BulletHandler.getBulletItem(BulletItem.POTION))),
+					getFluidTagForType(potion, 250)
+			);
+		};
+
 		PotionHelper.applyToAllPotionRecipes((out, in, reagent) -> {
-					if(!recipes.containsKey(out))
-						recipes.put(out, toRecipe.apply(out));
+					if(!bottleRecipes.containsKey(out))
+						bottleRecipes.put(out, toBottleRecipe.apply(out));
+					if(!bulletRecipes.containsKey(out))
+						bulletRecipes.put(out, toBulletRecipe.apply(out));
 				}
 		);
-		recipes.put(Potions.WATER, toRecipe.apply(Potions.WATER));
+		bottleRecipes.put(Potions.WATER, toBottleRecipe.apply(Potions.WATER));
 		IELogger.logger.info(
-				"Recipes for potions: "+recipes.keySet().stream()
+				"Recipes for potions: "+bottleRecipes.keySet().stream()
 						.map(BuiltInRegistries.POTION::getKey)
 						.map(ResourceLocation::toString)
 						.collect(Collectors.joining(", "))
 		);
-		return new ArrayList<>(recipes.values());
+		List<BottlingMachineRecipe> ret = new ArrayList<>(bottleRecipes.values());
+		ret.addAll(bulletRecipes.values());
+		return ret;
 	}
 
 	public static void registerPotionRecipe(Potion output, Potion input, IngredientWithSize reagent, Map<Potion, List<MixerRecipe>> all)